+++ packages/react-router/src/errors.ts
@@ -3,6 +3,7 @@ export {
   isEmailLinkError,
   isKnownError,
   isMetamaskError,
+  isReverificationCancelledError,
   EmailLinkErrorCode,
   EmailLinkErrorCodeStatus,
 } from '@clerk/clerk-react/errors';